+++ src/lib/carbon/index.ts
@@ -1,4 +1,5 @@
 export * from "./calculator";
+export * from "./constants";
 export * from "./factors";
 export * from "./quiz";
 export * from "./types";